feat: configure hyprland and nixos for improved nvidia support
The changes include:

- Corrected monitor transform in nixos.conf.
- Added environment variables for nvidia in hyprland.conf.
- Disabled border size in hyprland.conf.
- Added a script to restart waybar.
- Enabled nvidia-vaapi-driver in configuration.nix.
- Added EGL vendor library filename in configuration.nix.
- Forced video drivers to nvidia in configuration.nix.
- Enabled open gpu kernel modules in configuration.nix.
- Removed gamescope session in mrzelee.nix.
- Added gperftools to steam package in mrzelee.nix.
- Disabled steam controller configuration in mrzelee.nix.
- Added xwayland force zero scaling in hyprland.conf.

+++ .config/hypr/scripts/restart-waybar.sh
@@ -2,9 +2,18 @@
 
 WAYBAR_CMD="waybar -c $HOME/.config/waybar/${HYPRLAND_HOST}_config.jsonc"
 
+while [[ -z "$HYPRLAND_INSTANCE_SIGNATURE" ]]; do
+  sleep 1
+done
+
+
 if pgrep -fx "$WAYBAR_CMD" >/dev/null; then
     # Waybar is running, so kill it
-    pkill -fx "$WAYBAR_CMD"
+    if [[ $# -gt 0 && "$1" == "set" ]]; then
+      :
+    else
+      pkill -fx "$WAYBAR_CMD"
+    fi
 else
     # Waybar is not running, so launch it
     $WAYBAR_CMD | tee -a /tmp/waybar.log &
